refactor gpu code to separate TU
