+++ experiment.h
@@ -1,17 +1,19 @@
 #pragma once
 
+#include <memory>
 #include <string>
 #include <string_view>
 #include <vector>
 
 #include <fmt/core.h>
 
+using value_type=double;
+
 // there are three workloads that can be run:
 //  none:    do nothing
 //  dgemm:   run dgemm kernels full speed
 //  stream:  run dgemm kernels full speed
 enum class benchmark_kind {none, gemm, stream};
-
 
 static const char* benchmark_string(benchmark_kind k) {
     if      (k==benchmark_kind::none)   return "none";
@@ -26,6 +28,24 @@ struct experiment {
     benchmark_kind kind = benchmark_kind::none;
     std::vector<uint64_t> args;
 };
+
+
+struct benchmark {
+    virtual void run() = 0;
+    virtual void synchronize() = 0;
+    virtual std::string report(std::vector<double>) = 0;
+    virtual ~benchmark() {};
+};
+
+struct null_benchmark: benchmark {
+    null_benchmark(std::uint32_t) {}
+    void run() {};
+    void synchronize() {};
+    std::string report(std::vector<double>) {return "no benchmark run";};
+};
+
+std::unique_ptr<benchmark> get_gpu_benchmark(std::uint32_t N, benchmark_kind kind);
+std::unique_ptr<benchmark> get_cpu_benchmark(std::uint32_t N, benchmark_kind kind);
 
 // fmt library gubbins.
 
@@ -56,3 +76,6 @@ auto fmt::formatter<experiment>::format(experiment const& e, FormatContext& ctx)
     }
     return fmt::format_to(ctx.out(), "stream triad");
 }
+
+std::string flop_report_gemm(uint32_t N, std::vector<double> times);
+std::string bandwidth_report_stream(uint64_t N, std::vector<double> times);
